Let route- and rule- files contain shell variables.
This is necessary to handle dynamic assignments.  We cannot add a
static route to a table if the router IP changes.  This lets a route-
file contain something like:

    # The basic route for the interface:
    ${new_network_number}/${new_subnet_cidr} dev ${interface} table 2

    # The default route using the interface
    default via ${new_routers} dev ${interface} metric 1000 table 2

and the rule- file:

    from ${new_network_number}/${new_subnet_cidr}   lookup 2
    fwmark 2                                        lookup 2

This solves the problem of use of a DHCP-assigned IP in an
environment where some type of nontrivial/policy routing is
in use.